pyproject.toml
[project]
name = "lake"
version = "0.1.0"
description = "Data Lake initiator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.38.32",
    "duckdb==1.3.0",
    "kafka-python>=2.2.15",
    "numpy==2.2.6",
    "pandas==2.2.3",
    "psycopg2-binary==2.9.10",
    "pyarrow==20.0.0",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "pytest==8.3.5",
    "python-dotenv==1.1.0",
    "pyyaml>=6.0.2",
]



[project.scripts]
lake = "lake.cmd:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
